// cloudapp/src/app/models/response.ts
export class PriceResponse {
    requestId: string = "";
    status: string = "";
    errorMsg: string = "";
    price: string = "";
  }

  export class PlaceOrderResponse {
    requestId: string = "";
    status: string = "";
    errorMsg: string = "";
    randomNumber: string = "";
    additionalId: string = "";
  }